Take method failures into account
A method fails whenever there is no suitable receiver for a message within the
method. The method analysis is not aborted at that point, but continues normally
in order to collect as much type information as possible.
The method failures are relevant to determine duck types of a variable, every
duck type that failed a message with the variable as receiver, is removed.